Rename sti to page. Remove old page representation.
