+++ src/main/java/net/eugenpaul/jlexi/component/text/format/structure/TextSectionConfiguration.java
@@ -17,17 +17,17 @@ public class TextSectionConfiguration {
     @Default
     private int numberOfColumns = 1;
     @Default
-    private int siteWidthPx = 595;
+    private int pageWidthPx = 595;
     @Default
-    private int siteHeightPx = 842;
+    private int pageHeightPx = 842;
     @Default
     private boolean block = false;
 
     public TextSectionConfiguration copy() {
         return new TextSectionConfiguration(//
                 this.numberOfColumns, //
-                this.siteWidthPx, //
-                this.siteHeightPx, //
+                this.pageWidthPx, //
+                this.pageHeightPx, //
                 this.block //
         );
     }
